tissot: readmac: add support for reading bluetooth mac
The QTI bluetooth HAL should read the mac from persist but fails
to do so. As a fallback, it checks for a randomly generated
mac from /data/misc/bluedroid/btaddr_random. Store the proper
mac read using libqminvapi in that file to make the bluetooth HAL
read the proper mac and handle random mac generation if
libqminvapi is not available.
chown and chmod the path to allow the HAL to load it.

Noop changes:
 * rename mac file defines for uniformity
 * rename mac size define
 * rename libqminvapi function type to a more generic name
 * move random mac generation to a separate function
 * move mac loading to a separate function and pass
   libqminvapi mac loading function in a parameter
 * make the wlan mac write function return unix-style

Change-Id: I401208389ee49e16eb3a8ab62653d1625a4e90b4